migration tasks to ts

+++ src/resources/tasks/task.router.ts
@@ -1,53 +1,70 @@
-const router = require('express').Router({ mergeParams: true });
-const Task = require('./task.model');
-const taskService = require('./task.service');
+import { Request, Response, Router } from 'express';
+import taskService from './task.service';
+import Task from './task.model';
 
-router.route('/').get(async (req, res) => {
+const router = Router({ mergeParams: true });
+
+router.route('/').get(async (req: Request, res: Response) => {
   try {
-    const { boardId } = req.params
-    const tasks = await taskService.getAll(boardId);
-    res.json(tasks.map(Task.toResponse));
+    const { boardId } = req.params;
+    if (typeof boardId === 'string') {
+      const tasks = await taskService.getAll(boardId);
+      res.json(tasks.map(Task.toResponse));
+    }
+
   } catch ({ message }) {
     res.status(404).send(message);
   }
 });
 
-router.route('/:taskId').get(async (req, res) => {
+router.route('/:taskId').get(async (req: Request, res: Response) => {
   try {
     const { boardId, taskId } = req.params;
-    const task = await taskService.getById(boardId, taskId);
-    res.json(Task.toResponse(task));
+    if (typeof boardId === 'string' && taskId === 'string') {
+      const task = await taskService.getById(boardId, taskId);
+      res.json(Task.toResponse(task));
+    }
+
   } catch ({ message }) {
     res.status(404).send(message);
   }
 });
 
-router.route('/').post(async (req, res) => {
+router.route('/').post(async (req: Request, res: Response) => {
   try {
     const { boardId } = req.params;
     const newTask = new Task({ ...req.body })
-    const task = await taskService.create(boardId, { ...newTask, boardId });
-    res.status(201).json(Task.toResponse(task));
+    if (typeof boardId === 'string') {
+      const task = await taskService.create(boardId, { ...newTask, boardId });
+      res.status(201).json(Task.toResponse(task));
+    }
+
   } catch ({ message }) {
     res.status(404).send(message);
   }
 });
 
-router.route('/:taskdId').put(async (req, res) => {
-  try {
-    const { boardId, taskdId } = req.params;
-    const user = await taskService.update(boardId, taskdId, req.body);
-    res.json(Task.toResponse(user));
-  } catch ({ message }) {
-    res.status(404).send(message);
-  }
-});
-
-router.route('/:taskId').delete(async (req, res) => {
+router.route('/:taskdId').put(async (req: Request, res: Response) => {
   try {
     const { boardId, taskId } = req.params;
-    await taskService.remove(boardId, taskId);
-    res.status(204).json(`Task is deleted with id = ${taskId}`);
+    if (typeof boardId === 'string' && taskId === 'string') {
+      const user = await taskService.update(boardId, taskId, req.body);
+      res.json(Task.toResponse(user));
+    }
+
+  } catch ({ message }) {
+    res.status(404).send(message);
+  }
+});
+
+router.route('/:taskId').delete(async (req: Request, res: Response) => {
+  try {
+    const { boardId, taskId } = req.params;
+    if (typeof boardId === 'string' && taskId === 'string') {
+      await taskService.remove(boardId, taskId);
+      res.status(204).json(`Task is deleted with id = ${taskId}`);
+    }
+
   } catch ({ message }) {
     res.status(404).send(message);
   }
